Correctly configure the package for Clear Linux
This change ensures we differentiate between emul32 and native requirements,
whilst restoring all files except gtk-doc, for usage within the desktop
bundle (GNOME).

The package is modified to now operate in the absence of configuration,
as well as focusing solely on non-ethernet connections, which has now been
explicitly disabled with the included patch.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>